Display list of sub events on event page
