+++ src/domain/event/eventInfo/otherEventTimes.module.scss
@@ -1,0 +1,53 @@
+@import "colors";
+
+.otherEventTimes {
+  margin-bottom: 1rem;
+
+  .toggleButton {
+    padding: 0.75rem 0;
+    font-weight: 600;
+    display: flex;
+    width: 100%;
+    align-items: center;
+
+    &:focus {
+      outline: 2px solid $hel-black;
+    }
+
+    span {
+      flex: 1 1 0%;
+      text-align: left;
+    }
+
+    svg {
+      height: 1rem;
+      width: 1rem;
+    }
+
+    &.isListOpen {
+      svg {
+        transform: rotateX(180deg);
+      }
+    }
+  }
+
+  ul.timeList {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+
+    li {
+      display: flex;
+      padding: 0.75rem 0;
+      border-top: 1px solid $hel-black-10;
+
+      &:last-child {
+        border-bottom: 1px solid $hel-black-10;
+      }
+
+      span {
+        flex: 1 1 0%;
+      }
+    }
+  }
+}
